Restore FIELD actor depth ordering and source interpolation
Coalesce paired local OT buckets into the shared world ordering table, preserve source FIFO at interpolated midpoints, and derive moving depth from captured source transforms. Use owned scene projection for shadows and widen world-sector admission with the projection.

Wire Spyro's per-instance temporal scene source and evidence accounting. Make launcher configuration use Ninja without a duplicate compiler-selection stamp, preserve nested builds during migration, and record the independent-console observation and remaining timing/release gaps.

Validated with Clang/Ninja, clang-tidy across 114 translation units, 203 format checks, all 15 CTests, and a zero-compilation unchanged build. Artisans observation: 7,659,432 JIT blocks, zero fallback, 833 midpoint/endpoint pairs; full visual/timing parity remains open.

+++ game/render/paired_actor_decode.cpp
@@ -369,7 +369,7 @@ FaceCompareResult compare_ordered_faces(std::span<const ResolvedFace> expected,
 static ResolveResult resolve_normal_faces_impl(std::span<const Primitive> primitives,
                                                std::span<const ProjectedVertex> projected,
                                                const MaterialTables &materials,
-                                               uint32_t depth_origin,
+                                               double depth_origin,
                                                uint8_t shift,
                                                bool continuous) {
   ResolveResult result;
@@ -444,7 +444,12 @@ static ResolveResult resolve_normal_faces_impl(std::span<const Primitive> primit
       face.continuous_ot_key = raw / scale;
       face.ot_raw = (uint32_t)std::clamp(raw, 0.0, (double)UINT32_MAX);
       face.ot_bin = (uint32_t)std::clamp(face.continuous_ot_key, 0.0, (double)UINT32_MAX);
-    } else if (!compute_ot_bin(emitted, depth, depth_origin, shift, face.ot_raw, face.ot_bin)) {
+    } else if (!compute_ot_bin(emitted,
+                               depth,
+                               static_cast<uint32_t>(depth_origin),
+                               shift,
+                               face.ot_raw,
+                               face.ot_bin)) {
       continue;
     }
     face.quad ? ++result.quads : ++result.triangles;
@@ -470,7 +475,7 @@ ResolveResult resolve_normal_faces(std::span<const Primitive> primitives,
 ResolveResult resolve_normal_faces_continuous(std::span<const Primitive> primitives,
                                               std::span<const ProjectedVertex> projected,
                                               const MaterialTables &materials,
-                                              uint32_t depth_origin,
+                                              double depth_origin,
                                               uint8_t shift) {
   return resolve_normal_faces_impl(primitives, projected, materials, depth_origin, shift, true);
 }
